Adjust pmix_hotel_init eviction to 64 bit
 * pmix_hotel_init now uses a `uint64_t` instead of a `uint32_t`
   for the eviction value to allow for higher usec values.

+++ src/class/pmix_hotel.c
@@ -3,6 +3,7 @@
  * Copyright (c) 2012-2016 Cisco Systems, Inc.  All rights reserved.
  * Copyright (c) 2012      Los Alamos National Security, LLC. All rights reserved
  * Copyright (c) 2015-2018 Intel, Inc.  All rights reserved.
+ * Copyright (c) 2020      IBM Corporation.  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -48,7 +49,7 @@ static void local_eviction_callback(int fd, short flags, void *arg)
 
 pmix_status_t pmix_hotel_init(pmix_hotel_t *h, int num_rooms,
                               pmix_event_base_t *evbase,
-                              uint32_t eviction_timeout,
+                              uint64_t eviction_timeout,
                               pmix_hotel_eviction_callback_fn_t evict_callback_fn)
 {
     int i;
